L1-support-tools/01-ticket-triage-cli/macos/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ticket-triage"
version = "1.0.0"
description = "CLI tool for L1 support ticket triage"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "L1 Support Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = [
    "requests>=2.28",
]

[project.scripts]
ticket-triage = "ticket_triage.__main__:cli"

[project.urls]
Homepage = "https://github.com/your-org/ticket-triage-cli"

[tool.hatch.build.targets.wheel]
packages = ["ticket_triage"]

[tool.hatch.build.targets.wheel.force-include]
"ticket_triage/rules.yaml" = "ticket_triage/rules.yaml"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
